4.0.0/stability - Move Delete dialog to the components directory, delete empty directories that will not be used
